+++ QiuHRenderer/src/Support/Math/Triangle.h
@@ -3,4 +3,4 @@
 #include "Support/Math/Math.h"
 
 std::tuple<Vec2i, Vec2i> GetBoudingBox(Vec3f* vertex);
-Vec3f GetBarycentricCoordinate(Vec3f* triangle, Vec2i targetFragment);
+Vec3f GetBarycentricCoordinate(Vec3f* triangle, Vec2f targetFragment);
